Clean up TokenAuth creation and token saving; update README
